Implement various list data structures with exception handling
