+++ MyList.java
@@ -12,13 +12,3 @@ public interface MyList<E> {
     public boolean delete(E data); // returns false if the data is not deleted in the list
     public int search(E data); // returns index of data, else -1 is return
 }
-// File: ListOverflowException.java
-package prelim;
-public class ListOverflowException extends Exception {
-    public ListOverflowException() {
-        super("List is full! Cannot insert more elements.");
-    }
-    public ListOverflowException(String message) {
-        super(message);
-    }
-}
